Move VisitDataSource to separate file and fix linting errors
